version_1 and version_2
in version_2 we design the machine learning model after removing the variables 'specials' and'refuel_gas_type', as we considered these two variables already contained information provided in the dataset. we initially tried to do it with the 'drop' method applied to the dataset but the program did not properly recognise the new dataset after removing the variables, so we operated directly on the source data (.csv) by removing these two columns. 

important to say as well that the original dataset displayed decimals with commas (,) so we changed these values to dot (.) so that the program could properly read them. 

in both files we perform the entire exploratory data analysis (eda), including histogram, outliers, correlations, and also the machine learning model, as well as hypothesis testing with anova. next step consists of the streamlit user interface. 

we also attach the data visualisation panels from tableau, available in this link: https://public.tableau.com/views/CarFuelConsumption/Dashboard1?:language=es-ES&:display_count=n&:origin=viz_share_link