Add files via upload
we have two versions of the jupyter notebook files, in the original one we kept the column 'specials', which makes reference to the different driving conditions (snow, sun, rain, ac and the combinations with each other, it was though a bit confusing as it was not displayed in a manageable way, to avoid confusions, reason why we decided to remove this column in the version 2 of the jynb file). We also suggested the python script for deploying the streamlit application to calculate the fuel consumption using as input variables the ones provided in the dataset. All relevant findings are summarised in the presentation provided. 

Some external data was used, obtained from the Ministry of Industry regarding the historical fuel prices as well as the most recent fuel prices in the Spanish market (source, https://www.clickgasoil.com/c/precio-gasolina-95).